+++ src/Bill.kt
@@ -2,7 +2,7 @@ import java.util.UUID
 
 class Bill(
     val payer: Person,
-    val amount: Int,          // in cents
+    val amount: Long,          // in cents
     val debtors: List<Person>
 ) {
     val id: UUID = UUID.randomUUID()
